requirements.txt
django==4.2.16
python-decouple==3.8
psycopg2-binary==2.9.10
django-image-cropping==1.7
easy_thumbnails==2.10
